rptest: Extract transactional producer to si_tools
